moved and renamed data-info.ts
