+++ src/app/data-loader/info/data-info.ts
@@ -1,8 +1,9 @@
-import { WorkspaceInfo } from './workspace-info';
+import { WorkspaceInfo } from '../workspace-info';
 export class DataInfo {
     // #region [Constants]
     static MILLISECONDS = 1000;
     // #endregion
+
     // #region [Properties]
     workspace: string;
     name: string;
@@ -17,6 +18,7 @@ export class DataInfo {
     crop?: number[];
     private ws: WorkspaceInfo;
     // #endregion
+
     // #region [Constructor]
     constructor(ws, info) {
         // provided properties
@@ -37,7 +39,9 @@ export class DataInfo {
             this.crop = info.crop;
     }
     // #endregion
+    
     // #region [Accessors]
     get rate() { return DataInfo.MILLISECONDS / this.Hz; }
     get isEnergy() { return this.type === 'energy'; }
+    // #endregion
 }
