+++ src/app/dataview/dataview.component.ts
@@ -4,7 +4,7 @@ import { Component, OnInit, HostListener, ViewChildren, ViewChild,
          ChangeDetectorRef, AfterViewChecked } from '@angular/core';
 import { DataloaderService } from '../data-loader/data-loader.service';
 import { WorkspaceInfo, TypeMap, LabelScheme } from '../data-loader/workspace-info';
-import { DataInfo } from "../data-loader/DataInfo";
+import { DataInfo } from "../data-loader/info/data-info";
 import { SettingsService } from '../settings/settings.service';
 import { Label, LabelStream} from './labelstreams/labelstream';
 import { ModeTracker } from './modes/tool-mode';
